Adição de arquivos

+++ Aula 01/exercicio2.py
@@ -1,6 +1,8 @@
-a = int(input("Digite o primeiro número: "))
-b = int(input("Digite o segundo número: "))
+# Input: Dois números reais
+a = float(input("Digite o primeiro número: "))
+b = float(input("Digite o segundo número: "))
 
+# Condição para saber qual número é o maior
 if a < b:
     print("O menor número é ", a)
 else:
